Midterm

+++ src/design/Employee.java
@@ -18,7 +18,9 @@ public interface Employee {
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
+
+
 	
 	//employee benefit
 	public void benefitLayout();
